% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgo_api_ach.R
\name{csgo_api_ach}
\alias{csgo_api_ach}
\title{CS Go Achievements}
\usage{
csgo_api_ach(api_key, user_id)
}
\arguments{
\item{api_key}{string with the key provided by the steam API.

PS: If you don`t have a API key yet go to  \url{https://steamcommunity.com/login/home/?goto=\%2Fdev\%2Fapikey} and follow the presented steps.}

\item{user_id}{string with the steam user ID.

Steam ID is the NUMBER OR NAME at the end of your steam profile URL. ex: '76561198263364899'.

PS: The user should have a public status.}
}
\value{
data frame with all the CS Go achievements of the user ID.
}
\description{
This function will return all the CS Go Achievements of the user_id (input).
}
\examples{
\dontrun{
## It is necessary to fill the "api_key" parameter to run the example

df_ach <- csgo_api_ach(api_key = 'XXX', user_id = '76561198263364899')
}
}
