% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgo_api_stats.R
\name{csgo_api_stats}
\alias{csgo_api_stats}
\title{CS Go Statistics}
\usage{
csgo_api_stats(api_key, user_id)
}
\arguments{
\item{api_key}{string with the key provided by the steam API.

PS: If you don`t have a API key yet run \code{vignette("auth", package = "CSGo")} and follow the presented steps.}

\item{user_id}{string with the steam user ID.

Steam ID is the NUMBER OR NAME at the end of your steam profile URL. ex: '76561198263364899'.

PS: The user should have a public status.}
}
\value{
data frame with all the CS Go statistics of the user ID.
}
\description{
This function will return all the CS Go Statistics of the user_id (input).
}
\examples{
\dontrun{
## It is necessary to fill the "api_key" parameter to run the example

df_stats <- csgo_api_stats(api_key = 'XXX', user_id = '76561198263364899')
}
}
