% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_pictures}
\alias{map_pictures}
\title{Maps Images}
\format{
A data frame with 34 rows and 2 variables:
\describe{
\item{map_name}{Name of the map.}
\item{map_photo}{The image address.}
...
}
}
\source{
Created by the author.
}
\usage{
map_pictures
}
\description{
A dataset containing the pictures of each map.
}
\keyword{datasets}
