% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suport}
\alias{suport}
\title{Categories and Descriptions of the Statistics Data}
\format{
A data frame with 133 rows and 4 variables:
\describe{
\item{name_match}{Name to match with the name statitics data.}
\item{category}{Category name of the statistic.}
\item{desc}{Statistic description.}
\item{type}{Statistic type.}
...
}
}
\source{
Created by the author.
}
\usage{
suport
}
\description{
A dataset containing the categories, descriptions and types of the statistics data pulled from the csgo_api_stats.
}
\keyword{datasets}
