% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weapon_pictures}
\alias{weapon_pictures}
\title{Weapon Images}
\format{
A data frame with 34 rows and 2 variables:
\describe{
\item{weapon_name}{Name of the weapon.}
\item{weapon_photo}{The image address.}
...
}
}
\source{
Created by the author.
}
\usage{
weapon_pictures
}
\description{
A dataset containing the pictures of each map.
}
\keyword{datasets}
